Introduce type NestableDeclSyntax and NestableTypeInfo.swift
fixpu
