+++ Sources/SwiftInspectorVisitors/EnumVisitor.swift
@@ -149,11 +149,3 @@ public final class EnumVisitor: SyntaxVisitor {
   private var enumInfo: EnumInfo?
   private var innerEnums = [EnumInfo]()
 }
-
-public struct EnumInfo: Codable, Equatable {
-  public let name: String
-  public let inheritsFromTypes: [TypeDescription]
-  public let parentType: TypeDescription?
-  public let modifiers: Set<String>
-  // TODO: also find and expose properties on a class
-}
